CBVCVB

+++ toishan/src/main/java/com/integrador/toishan/model/Venta.java
@@ -15,7 +15,7 @@ public class Venta {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idVenta;
+    private Long idVenta;
 
     // Relación con Cliente (Muchos pedidos pertenecen a un cliente)
     @ManyToOne
@@ -38,11 +38,11 @@ public class Venta {
     @OneToMany(mappedBy = "venta", cascade = CascadeType.ALL)
     private List<DetalleVenta> detalles;
 
-    public Integer getIdVenta() {
+    public Long getIdVenta() {
         return idVenta;
     }
 
-    public void setIdVenta(Integer idVenta) {
+    public void setIdVenta(Long idVenta) {
         this.idVenta = idVenta;
     }
